% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_create_workspace.R
\name{pbi_create_workspace}
\alias{pbi_create_workspace}
\title{Create a workspace with custom parameters}
\usage{
pbi_create_workspace(name, key)
}
\arguments{
\item{name}{The name of the new workspace}

\item{key}{PowerBI application's Azure token}
}
\value{
HTTP code to know the good/bad state of the POST request
}
\description{
Create a workspace with custom parameters
}
