% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_export_report.R
\name{pbi_export_report}
\alias{pbi_export_report}
\title{Export a report from a given workspace}
\usage{
pbi_export_report(wsId, key, reportId, name)
}
\arguments{
\item{wsId}{Workspace's Id}

\item{key}{PowerBI application's Azure token}

\item{reportId}{Report Id from the report which you want to export}

\item{name}{The name of the exported report}
}
\value{
HTTP code to know the good/bad state of the POST request
}
\description{
Export a report from a given workspace
}
