% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_get_azure_token.R
\name{pbi_get_azure_token}
\alias{pbi_get_azure_token}
\title{retrieve Azure token for PBI app}
\usage{
pbi_get_azure_token(tenant, app)
}
\arguments{
\item{tenant}{Directory (tenant) ID}

\item{app}{Application (client) ID}
}
\value{
The Azure toke for your PowerBI App
}
\description{
retrieve Azure token for PBI app
}
