% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_get_dataset_id.R
\name{pbi_get_dataset_id}
\alias{pbi_get_dataset_id}
\title{Get dataset id from a given workspace}
\usage{
pbi_get_dataset_id(wsId, key)
}
\arguments{
\item{wsId}{Workspace's target Id}

\item{key}{PowerBI application's Azure token}
}
\value{
The Id from the first dataset inside the workspace which id you have passed as parameter
}
\description{
Get dataset id from a given workspace
}
