% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_get_report_id_in_workspace.R
\name{pbi_get_report_id_in_workspace}
\alias{pbi_get_report_id_in_workspace}
\title{Get workspace id through it's name}
\usage{
pbi_get_report_id_in_workspace(wsId, key)
}
\arguments{
\item{wsId}{Workspace's Id}

\item{key}{PowerBI application's Azure token}
}
\value{
The Id from the first report inside the workspace which id you have passed as parameter
}
\description{
Get workspace id through it's name
}
