% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_get_title_from_report.R
\name{pbi_get_title_from_report}
\alias{pbi_get_title_from_report}
\title{Get report title from a given workspace}
\usage{
pbi_get_title_from_report(wsId, reportId, key)
}
\arguments{
\item{wsId}{Workspace's target Id}

\item{reportId}{Report's target Id}

\item{key}{PowerBI application's Azure token}
}
\value{
The title from the report inside the workspace which id you have passed as parameter
}
\description{
Get report title from a given workspace
}
