% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_get_workspace_id_by_name.R
\name{pbi_get_workspace_id_by_name}
\alias{pbi_get_workspace_id_by_name}
\title{Get workspace id through it's name}
\usage{
pbi_get_workspace_id_by_name(name, key)
}
\arguments{
\item{name}{Workspace's name}

\item{key}{PowerBI application's Azure token}
}
\value{
The Id from the workspace which name you have passed as parameter
}
\description{
Get workspace id through it's name
}
