% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_import_report.R
\name{pbi_import_report}
\alias{pbi_import_report}
\title{Export a report from a given workspace}
\usage{
pbi_import_report(wsId, name, key)
}
\arguments{
\item{wsId}{Workspace target Id}

\item{name}{The name of the .pbix which you want to import}

\item{key}{PowerBI application's Azure token}
}
\value{
HTTP code to know the good/bad state of the POST request
}
\description{
Export a report from a given workspace
}
