% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_refresh_workspace.R
\name{pbi_refresh_workspace}
\alias{pbi_refresh_workspace}
\title{Refresh a workspace data}
\usage{
pbi_refresh_workspace(wsId, dsId, key)
}
\arguments{
\item{wsId}{Workspace's target Id}

\item{dsId}{Dataset's target ID}

\item{key}{PowerBI application's Azure token}
}
\value{
HTTP code to know the good/bad state of the POST request
}
\description{
Refresh a workspace data
}
